md, raid10, LLVMLinux: Remove VLAIS from raid10 driver
Replaced the use of a Variable Length Array In Struct (VLAIS) with a C99
compliant equivalent. This patch instead allocates the appropriate amount of
memory using a char array through a VLA which is compliant with C99 and gcc
extensions.

This could be also implemented using memory allocations on heap but the
original implementation was using stack allocation then it is safer to
keep this way.

The new code can be compiled with both gcc and clang.

Signed-off-by: Behan Webster <behanw@converseincode.com>
Signed-off-by: Vinícius Tinti <viniciustinti@gmail.com>
Suggested-by: Arnd Bergmann <arnd@arndb.de>
Cc: Arnd Bergmann <arnd@arndb.de>