fs, nfs, LLVMLinux: Remove VLAIS from nfs
Replaced the use of a Variable Length Array In Struct (VLAIS) with a C99
compliant equivalent. This patch instead makes a zero size array at the
end of helper struct and delegates the allocation of the variable length
items to kzalloc.

The new code can be compiled with both gcc and clang.

Signed-off-by: Behan Webster <behanw@converseincode.com>
Reviewed-by: Vinícius Tinti <viniciustinti@gmail.com>
Suggested-by: Arnd Bergmann <arnd@arndb.de>
Cc: Arnd Bergmann <arnd@arndb.de>